requirements.txt
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
cv==1.0.0
distro==1.9.0
exceptiongroup==1.3.1
filelock==3.20.0
fsspec==2025.12.0
groq==0.37.1
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
pillow==12.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pygame-ce==2.5.6
requests==2.32.5
sniffio==1.3.1
sympy==1.14.0
torch==2.9.1
torchaudio==2.9.1
torchvision==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
